set field name to bigint

+++ migrations/frontend/1658122170_add_batch_change_id_to_batch_spec/up.sql
@@ -11,7 +11,7 @@
 --    in the associated metadata.yaml file.
 
 ALTER TABLE batch_specs
-    ADD COLUMN IF NOT EXISTS batch_change_id integer,
+    ADD COLUMN IF NOT EXISTS batch_change_id bigint,
     ADD FOREIGN KEY (batch_change_id) REFERENCES batch_changes(id) ON DELETE SET NULL DEFERRABLE;
 
 UPDATE batch_specs SET batch_change_id = (SELECT bc.id FROM batch_changes bc WHERE bc.id = batch_specs.id);
